% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terrorbar.R
\name{geom_terrorbar}
\alias{geom_terrorbar}
\title{Vertical Terrorbars}
\usage{
geom_terrorbar(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  na.rm = FALSE,
  orientation = NA,
  connect = TRUE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{ggplot2::aes()}}. If specified and \code{inherit.aes = TRUE} (the default), it is combined with the default mapping at the top level of the plot. You must supply mapping if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three options:
If \code{NULL}, the default, the data is inherited from the plot data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}.
A \code{data.frame}, or other object, will override the plot data. All objects will be fortified to produce a data frame. See \code{\link[ggplot2:fortify]{ggplot2::fortify()}} for which variables will be created.
A \code{function} will be called with a single argument, the plot data. The return value must be a \code{data.frame}, and will be used as the layer data. A \code{function} can be created from a formula (\emph{e.g.,} \code{~ head(.x, 10)}).}

\item{stat}{The statistical transformation to use on the data for this layer, either as a \verb{ggproto Geom} subclass or as a string naming the stat stripped of the \code{stat_} prefix (\emph{e.g.,} \code{"count"} rather than \code{"stat_count"})}

\item{position}{Position adjustment, either as a string naming the adjustment (\emph{e.g.,} \code{"jitter"} to use \code{position_jitter}), or the result of a call to a position adjustment function. Use the latter if you need to change the settings of the adjustment.}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{ggplot2::layer()}}. These are often aesthetics, used to set an aesthetic to a fixed value, like \code{colour = "red"} or \code{size = 3}. They may also be parameters to the paired geom/stat.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with a warning. If \code{TRUE}, missing values are silently removed.}

\item{orientation}{The orientation of the layer. The default (\code{NA}) automatically determines the orientation from the aesthetic mapping. In the rare event that this fails it can be given explicitly by setting orientation to either \code{"x"} or \code{"y"}. See the *Orientation` section for more detail.}

\item{connect}{logical. Whether or not the Ts should be connected with a line}

\item{show.legend}{logical. Should this layer be included in the legends? \code{NA}, the default, includes if any aesthetics are mapped. \code{FALSE} never includes, and \code{TRUE} always includes. It can also be a named logical vector to finely select the aesthetics to display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics, rather than combining with them. This is most useful for helper functions that define both data and aesthetics and shouldn't inherit behaviour from the default plot specification, \emph{e.g.,} \code{\link[ggplot2:borders]{ggplot2::borders()}}.}
}
\description{
Creates errorbars inspired by Figure 9 of the now retracted classic \emph{\href{https://www.hindawi.com/journals/amse/2022/3802603/}{"Monitoring of Sports Health Indicators Based on Wearable Nanobiosensors"}} (Gong & Liu, retracted). Unlike the original figure, graphics created with \code{geom_terrorbar} are actual errorbars.
}
\section{Orientation}{
This geom treats each axis differently and, thus, can thus have two orientations. Often the orientation is easy to deduce from a combination of the given mappings and the types of positional scales in use. Thus, ggplot2 will by default try to guess which orientation the layer should have. Under rare circumstances, the orientation is ambiguous and guessing may fail. In that case the orientation can be specified directly using the \code{orientation} parameter, which can be either \code{"x"} or \code{"y"}. The value gives the axis that the geom should run along, \code{"x"} being the default orientation you would expect for the geom.
}

\section{Aesthetics}{
\code{geom_terrorbar()} understands the following aesthetics (required aesthetics are in bold):
\itemize{
\item \strong{\code{x}} or \strong{\code{y}}
\item \strong{\code{ymin}} or \strong{\code{xmin}}
\item \strong{\code{ymax}} or \strong{\code{xmax}}
\item \code{alpha}
\item \code{colour}
\item \code{group}
\item \code{fontface} - one of \code{c("plain", "bold", "italic")}
\item \code{size} - size of the Ts in pt units
\item \code{line_type} - type of connecting line when \code{connect = TRUE}
\item \code{line_width} - width of connecting line when \code{connect = TRUE}
\item \code{line_alpha} - alpha for connecting line when \code{connect = TRUE}. By default = 0.6 * \code{alpha}
}
}

\seealso{
This is just a frivolous edit of \code{\link[ggplot2:geom_linerange]{ggplot2::geom_errorbar}}
}
